# src/presentation/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Speedia OS API",
        "title": "OsApi",
        "termsOfService": "https://speedia.net/tos/",
        "contact": {
            "name": "Speedia Engineering",
            "url": "https://speedia.net/",
            "email": "eng+swagger@speedia.net"
        },
        "license": {
            "name": "Eclipse Public License v2.0",
            "url": "https://www.eclipse.org/legal/epl-2.0/"
        },
        "version": "0.0.1"
    },
    "host": "localhost:1618",
    "basePath": "/_/api",
    "paths": {
...
        "/v1/database/{dbType}/": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List databases names, users and sizes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "GetDatabases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DatabaseType (like mysql, postgres)",
                        "name": "dbType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Database"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "CreateDatabase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DatabaseType (like mysql, postgres)",
                        "name": "dbType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "All props are required.",
                        "name": "createDatabaseDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDatabase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "DatabaseCreated",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/v1/database/{dbType}/{dbName}/": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "DeleteDatabase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DatabaseType (like mysql, postgres)",
                        "name": "dbType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "DatabaseName",
                        "name": "dbName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DatabaseDeleted",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/v1/database/{dbType}/{dbName}/user/": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new database user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "CreateDatabaseUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DatabaseType (like mysql, postgres)",
                        "name": "dbType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "DatabaseName",
                        "name": "dbName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "privileges is optional. When not provided, privileges will be 'ALL'.",
                        "name": "createDatabaseUserDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDatabaseUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "DatabaseUserCreated",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/v1/database/{dbType}/{dbName}/user/{dbUser}/": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a database user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "DeleteDatabaseUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DatabaseType (like mysql, postgres)",
                        "name": "dbType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "DatabaseName",
                        "name": "dbName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Database username that will be deleted.",
                        "name": "dbUser",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DatabaseUserDeleted",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateDatabase": {
            "type": "object",
            "properties": {
                "dbName": {
                    "type": "string"
                }
            }
        },
        "dto.CreateDatabaseUser": {
            "type": "object",
            "properties": {
                "dbName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "privileges": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "entity.Database": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DatabaseUser"
                    }
                }
            }
        },
        "entity.DatabaseUser": {
            "type": "object",
            "properties": {
                "dbName": {
                    "type": "string"
                },
                "dbType": {
                    "type": "string"
                },
                "privileges": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" + JWT token or API key.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}